Module dioxus_native_core::prelude

A prelude of commonly used items

Re-exports

pub use crate::node::ElementNode;

pub use crate::node::FromAnyValue;

pub use crate::node::NodeType;

pub use crate::node::OwnedAttributeView;

pub use crate::node::TextNode;

pub use crate::node_ref::AttributeMaskBuilder;

pub use crate::node_ref::NodeMaskBuilder;

pub use crate::node_ref::NodeView;

pub use crate::passes::run_pass;

pub use crate::passes::RunPassView;

pub use crate::real_dom::NodeImmutable;

pub use crate::real_dom::NodeMut;

pub use crate::real_dom::NodeRef;

pub use crate::real_dom::RealDom;

pub use crate::SendAnyMap;

pub use crate::dioxus::*;

Structs

DependancyView

A immutable view of a State

Dependants

The states that depend on this state

NodeId

Handle to an entity.

TypeErasedState

A type erased version of State that can be added to the crate::prelude::RealDom with crate::prelude::RealDom::new

Enums

PassDirection

The direction that a pass should be run in

Traits

Dependancy

A trait that is implemented for all the dependancies of a State

State

A state that is automatically inserted in a node with dependencies.