Trait dioxus_native_core::prelude::Dependancy

pub trait Dependancy {
    type ElementBorrowed<'a>;

    // Provided method
    fn type_ids() -> Box<[TypeId]> { ... }
}

A trait that is implemented for all the dependancies of a State

Required Associated Types

type ElementBorrowed<'a>

A tuple with all the elements of the dependancy as DependancyView

Provided Methods

fn type_ids() -> Box<[TypeId]>

Returns a list of all the TypeIds of the elements in the dependancy

Object Safety

This trait is not object safe.