Struct dioxus_native_core::utils::cursor::Pos

pub struct Pos {
    pub col: usize,
    pub row: usize,
}

A cursor position

Fields
col: usize

The virtual column of the cursor. This can be more than the line length. To get the realized column, use Pos::col().

row: usize

The row of the cursor.

Implementations

impl Pos

pub fn new(col: usize, row: usize) -> Self

Creates a new cursor position.

pub fn up(&mut self, text: &(impl Text + ?Sized))

Moves the position up by one line.

pub fn down(&mut self, text: &(impl Text + ?Sized))

Moves the position down by one line.

pub fn right(&mut self, text: &(impl Text + ?Sized))

Moves the position right by one character.

pub fn left(&mut self, text: &(impl Text + ?Sized))

Moves the position left by one character.

pub fn move_row(&mut self, change: i32, text: &(impl Text + ?Sized))

Move the position’s row by the given amount. (positive is down, negative is up)

pub fn move_col(&mut self, change: i32, text: &(impl Text + ?Sized))

Move the position’s column by the given amount. (positive is right, negative is left)

pub fn col(&self, text: &(impl Text + ?Sized)) -> usize

Get the realized column of the position. This is the column, but capped at the line length.

pub fn row(&self) -> usize

Get the row of the position.

pub fn idx(&self, text: &(impl Text + ?Sized)) -> usize

Get the character index of the position.

pub fn realize_col(&mut self, text: &(impl Text + ?Sized))

If the column is more than the line length, cap it to the line length.