Trait dioxus_native_core::utils::cursor::Text

pub trait Text {
    // Required methods
    fn line(&self, number: usize) -> Option<&Self>;
    fn length(&self) -> usize;
    fn line_count(&self) -> usize;
    fn character(&self, idx: usize) -> Option<char>;
    fn len_before_line(&self, line: usize) -> usize;
}

This contains the information about the text that is used by the cursor to handle navigation.

Required Methods

fn line(&self, number: usize) -> Option<&Self>

Returns the line at the given index.

fn length(&self) -> usize

Returns the length of the text in characters.

fn line_count(&self) -> usize

Returns the number of lines in the text.

fn character(&self, idx: usize) -> Option<char>

Returns the character at the given character index.

fn len_before_line(&self, line: usize) -> usize

Returns the length of the text before the given line in characters.

Object Safety

This trait is not object safe.